Update ESLint configuration to that needed by ESLint 9
The ultimate goal here is being able to use VS Code's ESLint extension,
even though our projects are not Node projects.

Previously, we used globally installed ESLint 8. With ESLint 9, which is
the current version, this no longer seems to be an option (and even with
version 8, this solution had its issues).

So now we'll try a different approach. From the point of view of ESLint,
the whole repository will be a single Node project, which will have
ESLint in its devDependencies. Users will have to once, at the
beginning, run `npm install`, even though we won't be otherwise using
npm at all.